Move s3 functionality from FileAndUrlUtis to S3Utils
